Refining the visuals of various components
Refining rendering of sliders
Refining rendering of progress bars
Remove all bytecode manipulation code and build tasks
Widgets for scroll bars, scroll panes and root panes
Color chooser starts
Color chooser shows hi DPI version of the crayons image
Moved hi DPI code to laf widget
Hi DPI aware icons for text component context menu entries
Hi DPI aware icons for auto scrolling of scroll panes
Convert preview of tabs, scroll pane and desktop frames to Hi DPI mode

+++ src/org/pushingpixels/substance/api/painter/border/CompositeBorderPainter.java
@@ -83,7 +83,7 @@ public class CompositeBorderPainter implements SubstanceBorderPainter {
 	}
 
 	@Override
-	public void paintBorder(Graphics g, Component c, int width, int height,
+	public void paintBorder(Graphics g, Component c, float width, float height,
 			Shape contour, Shape innerContour, SubstanceColorScheme borderScheme) {
 		if (innerContour != null) {
 			this.inner.paintBorder(g, c, width, height, innerContour, null,
