+++ src/org/pushingpixels/substance/api/painter/border/FractionBasedBorderPainter.java
@@ -74,7 +74,7 @@ public class FractionBasedBorderPainter extends FractionBasedPainter implements
 	}
 
 	@Override
-	public void paintBorder(Graphics g, Component c, int width, int height,
+	public void paintBorder(Graphics g, Component c, float width, float height,
 			Shape contour, Shape innerContour, SubstanceColorScheme borderScheme) {
 		if (contour == null)
 			return;
