+++ src/org/pushingpixels/substance/api/painter/border/StandardBorderPainter.java
@@ -78,7 +78,7 @@ public class StandardBorderPainter implements SubstanceBorderPainter {
 	}
 
 	@Override
-	public void paintBorder(Graphics g, Component c, int width, int height,
+	public void paintBorder(Graphics g, Component c, float width, float height,
 			Shape contour, Shape innerContour, SubstanceColorScheme borderScheme) {
 		if (contour == null)
 			return;
